+++ examples/05_device_functions.cpp
@@ -11,7 +11,7 @@ void axpy(int N, T alpha, const T* x, T* y) {
 }
 
 int main() {
-  const int vector_length = 2000;
+  const size_t vector_length = 2000;
   std::vector<float> x_host(vector_length, 1.0);
   std::vector<float> y_host(vector_length, 0.0);
   const float alpha = 1.0;
